moved everything in Server struct and cleaned main
